Add typescript to mobile app
