+++ mobile/package.json
@@ -13,7 +13,11 @@
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
     "react-native": "0.69.5",
-    "expo-notifications": "~0.16.1"
+    "expo-notifications": "~0.16.1",
+    "typescript": "^4.6.3",
+    "@types/react": "~18.0.0",
+    "@types/react-native": "~0.69.1",
+    "expo-device": "~4.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
